hello-api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
fastapi==0.115.11
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.14.0
idna==3.10
importlib_metadata==8.7.0
mypy==1.15.0
mypy-extensions==1.0.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-system-metrics==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
protobuf==5.29.5
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyroscope-io==0.8.11
pyroscope-otel==0.4.1
requests==2.32.4
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.23.0
